Switch to vitest in-test testing so we can test private things without breaching the public-api defined by package.json#exports
