+++ ember-scoped-css/src/lib/generateAbsolutePathHash.js
@@ -3,7 +3,7 @@ import path from 'node:path';
 
 import findUp from 'find-up';
 
-import generateHash from './generateRelativePathHash.js';
+import generateHash from './generateRelativePathHash.ts';
 
 export default function generateHashFromAbsolutePath(absolutePath) {
   /**
@@ -134,4 +134,26 @@ function workspacePackageName(sourcePath) {
   MANIFEST_CACHE.set(workspace, json);
 
   return json.name;
+}
+
+if (import.meta.vitest) {
+  const { it, expect } = import.meta.vitest;
+
+  it('should return a string', function () {
+    const postfix = generateHash('foo.css');
+
+    expect(postfix).to.be.a('string');
+  });
+
+  it('should return a string starting with "e"', function () {
+    const postfix = generateHash('foo.css');
+
+    expect(postfix).to.match(/^e/);
+  });
+
+  it('should return a string of length 9', function () {
+    const postfix = generateHash('foo.css');
+
+    expect(postfix).to.have.lengthOf(9);
+  });
 }
