+++ ember-scoped-css/src/template-lint/plugin.js
@@ -1,4 +1,4 @@
-import { Rule } from 'ember-template-lint';
+import { generateRuleTests, Rule } from 'ember-template-lint';
 
 class ScopedClassHelperRule extends Rule {
   visitor() {
@@ -31,9 +31,52 @@ class ScopedClassHelperRule extends Rule {
   }
 }
 
-export default {
+const scopedClassHelperPlugin = {
   name: 'scoped-css-plugin',
   rules: {
     'scoped-class-helper': ScopedClassHelperRule,
   },
 };
+
+export default scopedClassHelperPlugin;
+
+import assert from 'assert';
+
+if (import.meta.vitest) {
+  const { it, describe, beforeEach } = import.meta.vitest;
+
+  generateRuleTests({
+    name: 'scoped-class-helper',
+
+    groupMethodBefore: beforeEach,
+    groupingMethod: describe,
+    testMethod: it,
+    plugins: [scopedClassHelperPlugin],
+    config: true,
+    good: ['{{scoped-class "test"}}', '{{(scoped-class "test")}}'],
+    bad: [
+      {
+        template: '{{scoped-class}}',
+
+        verifyResults(results) {
+          assert.equal(results.length, 1);
+          assert.equal(
+            results[0].message,
+            'One positional param is required to be passed to scoped-class helper. {{scoped-class "some-class"}}. More info: https://github.com/soxhub/ember-scoped-css/blob/main/docs/lint-rules.md',
+          );
+        },
+      },
+      {
+        template: '{{scoped-class this.someClass}}',
+
+        verifyResults(results) {
+          assert.equal(results.length, 1);
+          assert.equal(
+            results[0].message,
+            'You cannot pass dynamic values to scoped-class helper. {{scoped-class "some-class"}}. More info: https://github.com/soxhub/ember-scoped-css/blob/main/docs/lint-rules.md',
+          );
+        },
+      },
+    ],
+  });
+}
