+++ ember-scoped-css/vitest.config.mts
@@ -1,0 +1,11 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    includeSource: ['src/**/*.{js,ts}'],
+  },
+  esbuild: {
+    target: 'esnext',
+    format: 'esm',
+  },
+});
